requirements.txt
openai>=1.30.0
python-dotenv>=1.0.1
trimesh>=4.4.9
numpy>=2.0.0
gmsh>=4.13.1